[WIP] Compiles but need to fix Consensus_test

+++ src/ripple/consensus/Consensus.h
@@ -916,7 +916,7 @@ Consensus<Derived, Traits>::getJson(bool full) const
     if (mode_ != Mode::wrongLedger)
     {
         ret["synched"] = true;
-        ret["ledger_seq"] = previousLedger_.seq() + 1;
+        ret["ledger_seq"] = static_cast<std::uint32_t>(previousLedger_.seq() + 1);
         ret["close_granularity"] = static_cast<Int>(closeResolution_.count());
     }
     else
